Add files via upload
This includes the basic theme (index in the main section), and the login page which is linked to in the navbar section.  This should work but I don't have time atm to fully test it- will do so later or tomorrow morning!  There is a lot more to be done obviously and on the theme side I've just changed the name titles, etc.  but it gives us a working basis.